Update css di Report

+++ pbw/pbw/src/main/java/com/tubes/Data/Show.java
@@ -14,7 +14,7 @@ public class Show {
     private Long idShow;           // ID untuk show
     private String showName;       // Nama show
     private Date date;             // Tanggal show
-    private int venue;          // Venue show
+    private Long venue;          // Venue id
     private String description;     // Deskripsi show
 }
